Create settings and move messages
